Generative AI/OCR App With Gemini and Streamlit/requirements.txt
Streamlit
PIL
google-generativeai
dotenv